feat: 标签，分类下有数据不能删除

+++ weblog-admin/src/main/java/com/arrebol/admin/service/impl/AdminTagServiceImpl.java
@@ -2,6 +2,10 @@ package com.arrebol.admin.service.impl;
 
 import com.arrebol.admin.model.vo.category.SelectRspVO;
 import com.arrebol.admin.model.vo.tag.*;
+import com.arrebol.common.domain.dos.ArticleTagRelDO;
+import com.arrebol.common.domain.mapper.ArticleTagRelMapper;
+import com.arrebol.common.exception.BizException;
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.arrebol.admin.service.AdminTagService;
@@ -18,6 +22,7 @@ import org.springframework.util.CollectionUtils;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,6 +31,8 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
 
     @Autowired
     private TagMapper tagMapper;
+    @Autowired
+    private ArticleTagRelMapper articleTagRelMapper;
 
     @Override
     public Response addTags(AddTagReqVO addTagReqVO) {
@@ -73,8 +80,17 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
     public Response deleteTag(DeleteTagReqVO deleteTagReqVO) {
         // 标签 ID
         Long tagId = deleteTagReqVO.getId();
+
+        // 校验该标签下是否有关联的文章，若有，则不允许删除，提示用户需要先删除标签下的文章
+        ArticleTagRelDO articleTagRelDO = articleTagRelMapper.selectOneByTagId(tagId);
+
+        if (Objects.nonNull(articleTagRelDO)) {
+            throw new BizException(ResponseCodeEnum.TAG_CAN_NOT_DELETE);
+        }
+
         // 根据标签 ID 删除
         int count = tagMapper.deleteById(tagId);
+
         return count == 1 ? Response.success() : Response.fail(ResponseCodeEnum.TAG_NOT_EXISTED);
     }
 
@@ -84,6 +100,23 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
         // 执行模糊查询
         List<TagDO> tagDOS = tagMapper.selectByKey(key);
         // do 转 vo
+        List<SelectRspVO> vos = null;
+        if (!CollectionUtils.isEmpty(tagDOS)) {
+            vos = tagDOS.stream()
+                    .map(tagDO -> SelectRspVO.builder()
+                            .label(tagDO.getName())
+                            .value(tagDO.getId())
+                            .build())
+                    .collect(Collectors.toList());
+        }
+        return Response.success(vos);
+    }
+
+    @Override
+    public Response findTagSelectList() {
+        // 查询所有标签
+        List<TagDO> tagDOS = tagMapper.selectList(Wrappers.emptyWrapper());
+        // DO 转 VO
         List<SelectRspVO> vos = null;
         if (!CollectionUtils.isEmpty(tagDOS)) {
             vos = tagDOS.stream()
